# apps/vuln-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title:  TestBed API (api-demo)
  version: 1.0.0
  description: >-
    Static OpenAPI spec for the api-demo app. For full narrative docs see
    API-DOCUMENTATION.md. Endpoints are intentionally vulnerable for testing.
servers:
  - url: http://localhost:8880
...
paths:
...
  /chain/{depth}:
    get:
      tags: [Testing]
      summary: Get
      parameters:
        - name: depth
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /error/{error_type}:
    get:
      tags: [Testing]
      summary: Get
      parameters:
        - name: error_type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      tags: [Testing]
      summary: Post
      parameters:
        - name: error_type
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /leak/headers:
    get:
      tags: [Testing]
      summary: Get
      responses:
        '200':
          description: OK
    post:
      tags: [Testing]
      summary: Post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /status/{code}:
    get:
      tags: [Testing]
      summary: Get
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    post:
      tags: [Testing]
      summary: Post
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...